__pycache__/requirements.txt
Flask==2.2.3
folium==0.14.0
pycountry==22.3.5
python-dotenv==1.0.0
requests==2.28.2
